+++ src/components/layouts/index.ts
@@ -1,4 +1,5 @@
 export * from './CardLayout'
+export * from './CardsStackLayout'
 export * from './LandingCardBase'
 export * from './PageLayout'
 export * from './SkeletonLayout'