+++ src/components/molecules/projectDisplay/RewardCard.tsx
@@ -93,7 +93,6 @@ export const RewardCard = ({
         </HStack>
         <Box>
           <ImageWithReload
-            grey
             borderRadius="4px"
             src={reward.image || ''}
             width="335px"