+++ src/context/filter.tsx
@@ -50,6 +50,8 @@ const defaultFilterContext = {
 
 export const FilterContext = createContext<FilterState>(defaultFilterContext)
 
+export const useFilterContext = () => useContext(FilterContext)
+
 export const FilterProvider = ({
   children,
   isLoggedIn,
@@ -65,54 +67,55 @@ export const FilterProvider = ({
   const navigate = useNavigate()
   const location = useLocation()
 
-  const getFiltersFromUrlParams = useCallback(() => {
-    const countryCode = searchParams.get('countryCode') || undefined
-    const region = searchParams.get('region') || undefined
-    const search = searchParams.get('search') || undefined
-    const status = (searchParams.get('status') as ProjectStatus) || undefined
-    const type = (searchParams.get('type') as ProjectType) || undefined
-    const tagIds =
-      searchParams
-        .get('tagIds')
-        ?.split(',')
-        ?.map((val) => toInt(val)) || undefined
-    const recent = Boolean(searchParams.get('recent')) || undefined
-    const activity =
-      (searchParams.get('activity') as ActivityResourceType) || undefined
-    const sort = (searchParams.get('sort') as SortType) || undefined
+  const updateFilter = useCallback(
+    async (value: Partial<FilterType>) => {
+      if (isLandingFeedPage && !isLoggedIn) {
+        navigate('/', { state: { save: true } })
+      }
 
-    return {
-      countryCode,
-      region,
-      search,
-      status,
-      type,
-      tagIds,
-      recent,
-      activity,
-      sort,
-    }
-  }, [searchParams])
+      const currentFilters = getFiltersFromUrlParams(searchParams)
 
-  const setUrlParamsFromFilters = useCallback(
-    (value: FilterType) => {
+      let newFilters = {} as FilterType
+
+      if (checkIfRenderFilter(value)) {
+        if (!currentFilters.sort && !isLandingFeedPage) {
+          newFilters = {
+            ...currentFilters,
+            recent: undefined,
+            sort: SortType.recent,
+            ...value,
+          }
+        } else {
+          newFilters = {
+            ...currentFilters,
+            recent: undefined,
+            ...value,
+          }
+        }
+      } else {
+        newFilters = {
+          ...currentFilters,
+          ...value,
+        }
+      }
+
       const newParameters = [] as [string, string][]
 
-      for (const key of Object.keys(value)) {
-        if (value[key as keyof FilterType]) {
+      for (const key of Object.keys(newFilters)) {
+        if (newFilters[key as keyof FilterType]) {
           if (key === 'tagIds') {
-            const tagIds = value[key as keyof FilterType] as number[]
+            const tagIds = newFilters[key as keyof FilterType] as number[]
             if (tagIds.length > 0) {
               newParameters.push([key, tagIds.join(',')])
             }
           } else if (
             key === 'sort' &&
-            value[key as keyof FilterType] === SortType.recent &&
-            disableSortByTrending(value)
+            newFilters[key as keyof FilterType] === SortType.recent &&
+            disableSortByTrending(newFilters)
           ) {
             newParameters.push([key, SortType.createdAt])
           } else {
-            newParameters.push([key, `${value[key as keyof FilterType]}`])
+            newParameters.push([key, `${newFilters[key as keyof FilterType]}`])
           }
         }
       }
@@ -120,65 +123,29 @@ export const FilterProvider = ({
       navigate(location.pathname, { state: null })
       setSearchParams(newParameters)
     },
-    [location, navigate, setSearchParams],
-  )
-
-  const updateFilter = useCallback(
-    async (value: Partial<FilterType>) => {
-      if (isLandingFeedPage && !isLoggedIn) {
-        navigate('/', { state: { save: true } })
-      }
-
-      const currentFilters = getFiltersFromUrlParams()
-
-      let newfilters = {} as FilterType
-
-      if (checkIfRenderFilter(value)) {
-        if (!currentFilters.sort && !isLandingFeedPage) {
-          newfilters = {
-            ...currentFilters,
-            recent: undefined,
-            sort: SortType.recent,
-            ...value,
-          }
-        } else {
-          newfilters = {
-            ...currentFilters,
-            recent: undefined,
-            ...value,
-          }
-        }
-      } else {
-        newfilters = {
-          ...currentFilters,
-          ...value,
-        }
-      }
-
-      setUrlParamsFromFilters(newfilters)
-    },
     [
       isLandingFeedPage,
-      getFiltersFromUrlParams,
-      setUrlParamsFromFilters,
-      navigate,
       isLoggedIn,
+      searchParams,
+      navigate,
+      location.pathname,
+      setSearchParams,
     ],
   )
 
   useEffect(() => {
-    const urlFilters = getFiltersFromUrlParams()
+    const urlFilters = getFiltersFromUrlParams(searchParams)
     setFilters(urlFilters)
     if (location.state?.save) {
       navigate('', { state: null })
     } else if (location.state?.filter) {
       updateFilter(location.state.filter)
     }
-  }, [location, getFiltersFromUrlParams, navigate, updateFilter])
+  }, [location, navigate, updateFilter, searchParams])
 
   const clearFilter = useCallback(() => {
-    setUrlParamsFromFilters({} as FilterType)
-  }, [setUrlParamsFromFilters])
+    setSearchParams({})
+  }, [setSearchParams])
 
   return (
     <FilterContext.Provider
@@ -193,4 +160,31 @@ export const FilterProvider = ({
   )
 }
 
-export const useFilterContext = () => useContext(FilterContext)
+const getFiltersFromUrlParams = (searchParams: URLSearchParams) => {
+  const countryCode = searchParams.get('countryCode') || undefined
+  const region = searchParams.get('region') || undefined
+  const search = searchParams.get('search') || undefined
+  const status = (searchParams.get('status') as ProjectStatus) || undefined
+  const type = (searchParams.get('type') as ProjectType) || undefined
+  const tagIds =
+    searchParams
+      .get('tagIds')
+      ?.split(',')
+      ?.map((val) => toInt(val)) || undefined
+  const recent = Boolean(searchParams.get('recent')) || undefined
+  const activity =
+    (searchParams.get('activity') as ActivityResourceType) || undefined
+  const sort = (searchParams.get('sort') as SortType) || undefined
+
+  return {
+    countryCode,
+    region,
+    search,
+    status,
+    type,
+    tagIds,
+    recent,
+    activity,
+    sort,
+  }
+}