+++ src/pages/projectView/projectActivityPanel/index.tsx
@@ -173,15 +173,7 @@ export const ProjectActivityPanel = ({ resourceType, resourceId }: Props) => {
 
     switch (fundState) {
       case fundingStages.initial:
-        return (
-          <ProjectFundingInitialInfoScreen
-            {...{
-              project,
-              fundingTx,
-              btcRate,
-            }}
-          />
-        )
+        return <ProjectFundingInitialInfoScreen project={project} />
       case fundingStages.form:
         return (
           <ProjectFundingSelectionFormScreen