+++ src/pages/projectView/projectActivityPanel/screens/ProjectFundingInitialInfoScreen.tsx
@@ -17,7 +17,7 @@ import {
 import { useQueryWithPagination } from '../../../../hooks'
 import { useFundSubscription } from '../../../../hooks/fundingFlow/useFundSubscription'
 import {
-  Funder,
+  FunderWithUserFragment,
   FundingTxFragment,
   ProjectFragment,
 } from '../../../../types/generated/graphql'
@@ -34,16 +34,11 @@ import { ProjectLeaderboardList } from '../components/ProjectLeaderboardList'
 
 type Props = {
   project: ProjectFragment
-  btcRate: number
-  fundingTx: FundingTxFragment
 }
 
 const itemLimit = 50
 
-export const ProjectFundingInitialInfoScreen = ({
-  project,
-  fundingTx,
-}: Props) => {
+export const ProjectFundingInitialInfoScreen = ({ project }: Props) => {
   const isMobile = useMobileMode()
   const { toast } = useNotification()
 
@@ -77,7 +72,7 @@ export const ProjectFundingInitialInfoScreen = ({
     },
   })
 
-  const funders = useQueryWithPagination<Funder>({
+  const funders = useQueryWithPagination<FunderWithUserFragment>({
     queryName: 'getFunders',
     itemLimit,
     query: QUERY_PROJECT_FUNDERS,
@@ -202,13 +197,13 @@ export const ProjectFundingInitialInfoScreen = ({
   return (
     <VStack
       padding={{ base: '0px 10px 0px 10px', lg: '10px 20px' }}
-      spacing="0px"
+      spacing={4}
       width="100%"
       height="100%"
       overflowY="hidden"
       position="relative"
     >
-      <ActivityBrief project={project} />
+      <ActivityBrief funders={funders.data} project={project} />
 
       {!isMobile ? (
         <Button
@@ -229,7 +224,6 @@ export const ProjectFundingInitialInfoScreen = ({
         alignItems="center"
         overflow="hidden"
         flex="1"
-        paddingTop="10px"
       >
         <Box display="flex" w="100%">
           {renderTabsList()}