+++ src/pages/projectView/projectMainBody/sections/index.ts
@@ -1,6 +1,7 @@
 export * from './CreatorSocial'
 export * from './Entries'
 export * from './Header'
+export * from './LaunchProjectNotice'
 export * from './Milestones'
 export * from './Rewards'
 export * from './Story'